requirements.txt
streamlit
tensorflow
opencv-python
pytesseract
numpy
Pillow